Allow use as application or library
Change the configuration and the code so the package can be used as
either a standalone application or as a library. Both Webpack config
files export a "library" and an "application" entrypoint. The dev
webpack.config.js is intended for when running the application
standalone and does not mark React as external while the prod
webpack.production.config.js marks React as external, relying on the
importing host library to provide the React libraries.

The entrypoint for the application is now `src/index.jsx` while the
entrypoint for the library is `src/lib-index.jsx`.
