+++ src/views/index.pug
@@ -5,4 +5,4 @@ block content
       window.__INITIAL_STATE__ = !{initialState}
   div(id='app') !{reactOutput}
 block bundle  
-  script(src='bundle.js')
+  script(src='application.bundle.js')
